Refactor: 각 폴더별 분류 및 정리
