+++ src/js/store/index.js
@@ -1,0 +1,10 @@
+const store = {
+  setLocalStorage(menu) {
+    localStorage.setItem('menu', JSON.stringify(menu));
+  },
+  getLocalStorage() {
+    return JSON.parse(localStorage.getItem('menu'));
+  }
+}
+
+export default store;
