+++ src/js/utils/dom.js
@@ -1,0 +1,1 @@
+export const $ = (selector) => document.querySelector(selector);
